Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "rukasuleonhart",email = "lucasoliveirasantossoares@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.1,<0.35.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "tzdata (>=2025.2,<2026.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
